requirements.txt
amqp==5.1.1
billiard==4.1.0
numpy==1.26.0
packaging==23.1
pandas==2.1.0
prompt-toolkit==3.0.39
python-certifi-win32==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.3
pytz==2023.3
redis==5.0.1
requests==2.31.0
setuptools-scm==7.1.0
six==1.16.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.8
wrapt==1.15.0
fastapi==0.111.0
uvicorn==0.30.1
Appium-Python-Client==2.11.1   
selenium==4.11.2
PyYAML==6.0.2
pytest==8.3.3